requirements.txt
    asgiref==3.5.0
    backports.zoneinfo==0.2.1;python_version<"3.9"

    certifi==2021.10.8
    charset-normalizer==2.0.12
    coreapi==2.3.3
    coreschema==0.0.4
    django-filter==21.1
    django-fsm==2.8.0
    django-jazzmin==2.6.0
    djangorestframework==3.13.1
    drf-yasg==1.20.0
    idna==3.3
    inflection==0.5.1
    itypes==1.2.0
    Jinja2==3.0.3
    MarkupSafe==2.1.0
    packaging==21.3
    
    pyparsing==3.0.7
    pytz==2021.3
    requests==2.27.1
    # ruamel.yaml
    # ruamel.yaml.clib
    Pillow
    sqlparse==0.4.2
    uritemplate==4.1.1
    urllib3==1.26.8
    # uWSGI>=2.0.19.1,<2.1
    django-cors-headers==3.11.0
    # psycopg-binary
    firebase-admin==5.2.0
    rsa
    djangorestframework-simplejwt
    psycopg2-binary
    openpyxl